+++ src/refactoring/features/coupon/components/Management.tsx
@@ -1,7 +1,6 @@
-import { useState } from 'react';
-import type { Coupon, DiscountType } from '../../../../types';
+import type { Coupon } from '../../../../types';
 import { CouponItem } from './Item';
-import { validateCoupon } from '../helpers';
+import { CouponForm } from './Form';
 
 interface CouponManagementProps {
   coupons: Coupon[];
@@ -9,91 +8,11 @@ interface CouponManagementProps {
 }
 
 export const CouponManagement = ({ coupons, onCouponAdd }: CouponManagementProps) => {
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: '',
-    code: '',
-    discountType: 'percentage',
-    discountValue: 0,
-  });
-  const [validationErrors, setValidationErrors] = useState<string[]>([]);
-
-  const handleAddCoupon = () => {
-    const validation = validateCoupon(newCoupon);
-    setValidationErrors(validation.errors);
-
-    if (!validation.isValid) {
-      return;
-    }
-
-    onCouponAdd(newCoupon);
-    setNewCoupon({
-      name: '',
-      code: '',
-      discountType: 'percentage',
-      discountValue: 0,
-    });
-    setValidationErrors([]);
-  };
-
   return (
     <div>
       <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
       <div className="bg-white p-4 rounded shadow">
-        <div className="space-y-2 mb-4">
-          <input
-            type="text"
-            placeholder="쿠폰 이름"
-            value={newCoupon.name}
-            onChange={(e) => {
-              setNewCoupon({ ...newCoupon, name: e.target.value });
-              setValidationErrors([]);
-            }}
-            className="w-full p-2 border rounded"
-          />
-          <input
-            type="text"
-            placeholder="쿠폰 코드"
-            value={newCoupon.code}
-            onChange={(e) => {
-              setNewCoupon({ ...newCoupon, code: e.target.value });
-              setValidationErrors([]);
-            }}
-            className="w-full p-2 border rounded"
-          />
-          <div className="flex gap-2">
-            <select
-              value={newCoupon.discountType}
-              onChange={(e) => {
-                setNewCoupon({ ...newCoupon, discountType: e.target.value as DiscountType });
-                setValidationErrors([]);
-              }}
-              className="w-full p-2 border rounded"
-            >
-              <option value="amount">금액(원)</option>
-              <option value="percentage">할인율(%)</option>
-            </select>
-            <input
-              type="number"
-              placeholder="할인 값"
-              value={newCoupon.discountValue}
-              onChange={(e) => {
-                setNewCoupon({ ...newCoupon, discountValue: parseInt(e.target.value) });
-                setValidationErrors([]);
-              }}
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <button onClick={handleAddCoupon} className="w-full bg-green-500 text-white p-2 rounded hover:bg-green-600">
-            쿠폰 추가
-          </button>
-          {validationErrors.length > 0 && (
-            <div className="mt-2 text-red-500 text-sm">
-              {validationErrors.map((error, index) => (
-                <div key={index}>{error}</div>
-              ))}
-            </div>
-          )}
-        </div>
+        <CouponForm onSubmit={onCouponAdd} />
         <div>
           <h3 className="text-lg font-semibold mb-2">현재 쿠폰 목록</h3>
           <div className="space-y-2">
